backend/django/yurnero/requirement.txt
pip>=10.0.0
setuptools>=39.0.1
wheel>=0.31.0
Django>=3.0
djangorestframework==3.10.3
django-rest-swagger==2.2.0
django-filter==2.2.0
django-cors-headers==3.2.0
python-redis-lock[django]==3.4.0
uWSGI==2.0.18
gunicorn==20.0.4


celery>=4.2.1
Pillow==>=5.1.0
sphinx>=1.7.2
pbr>=4.0.2
mysqlclient>=1.3.12
django-rest-auth>=0.9.3


psycopg2>=2.7.4
coreapi-cli>=1.0.9
gevent>=1.3.7
mysql-replication>=0.18
redis-collections>=0.5.2
django-redis>=4.10.0

pyzmq>=17.0.0
pika>=0.11.2
requests>=2.18.4
qrcode>=6.0

xlrd>=1.1.0
paho-mqtt>=1.4.0
msgpack>=0.6.1
django_silk>=3.0.1
openpyxl>=2.5.6
minio>=4.0.18
piexif>=1.1.2